aldizkariak/requirements.txt
marker-pdf
tqdm
requests
bs4
beautifulsoup4